+++ Log2Osvvm.tcl
@@ -18,6 +18,7 @@
 # 
 #  Revision History:
 #    Date      Version    Description
+#    07/2024   2024.07    Minor name updates
 #    06/2022   2022.06    Initial Revision
 #
 #