+++ VendorScripts_RivieraPro.tcl
@@ -19,6 +19,7 @@
 # 
 #  Revision History:
 #    Date      Version    Description
+#     7/2024   2024.07    Added DoWaves capability 
 #     5/2024   2024.05    Added CloseAllFiles to vendor_end_previous_simulation.  Added ToolVersion variable 
 #     5/2022   2022.05    Coverage report name based on TestCaseName rather than LibraryUnit
 #                         Updated variable naming 
@@ -193,6 +194,7 @@ proc vendor_simulate {LibraryName LibraryUnit args} {
   variable TestSuiteName
   variable TestCaseFileName
   variable SimulateOptions
+  variable WaveFiles
   global aldec            ; #  required for matlab cosim
 
   if {($::osvvm::Debug)} {
@@ -212,12 +214,30 @@ proc vendor_simulate {LibraryName LibraryUnit args} {
   if {$::osvvm::LogSignals} {
     log -rec [env]/*
   }
+  if {$WaveFiles ne ""} {
+    foreach wave $WaveFiles {
+      do $wave
+    }
+  }
+  set WaveFiles ""
   run -all 
   
   # Save Coverage Information 
   if {$::osvvm::CoverageEnable && $::osvvm::CoverageSimulateEnable} {
     acdb save -o ${::osvvm::CoverageDirectory}/${TestSuiteName}/${TestCaseFileName}.acdb -testname ${TestCaseFileName}
   }
+}
+
+# -------------------------------------------------
+proc vendor_DoWaves {args} {
+  variable WaveFiles
+  
+  if {$args ne ""} {
+    foreach wave {*}$args {
+      lappend WaveFiles $wave 
+    }
+  }
+  return ""
 }
 
 # -------------------------------------------------